refactor: split app.go by domain, lift the control dispatch out of serveControl
Two god-functions/files from the improvement plan, no behaviour change.

cmd/hivegui/app.go was 1101 lines and 50 methods. It already carried
section banners, so the split follows them: app_control.go (dial,
handshake, version banner, restart, control read loop), app_calls.go
(the thin RPCs — agents, sessions, projects, worktrees), app_attach.go
(per-session attach conns), app_spawn.go (dial-or-spawn hived). app.go
keeps the App type and lifecycle at 192 lines.

internal/daemon serveControl was 274 lines, the file's longest by 2x,
with a 14-case dispatch buried under four closures — and no way to
reach a handler without a live socket. The dispatch moves to
handleControlFrame(ctx, ops, ft, payload) bool, with the reply helpers
passed as a controlOps value; serveControl keeps the connection setup.
Twelve of the fourteen cases opened with the identical four-line
unmarshal-or-bad_payload prologue, now decodeReq[T].

The seam pays for itself immediately: control_frame_test.go drives the
handlers directly and pins what had no coverage — every decoding frame
answers bad_payload without dropping the client or touching the
registry, the list frames answer in place, an unknown frame is ignored
rather than fatal (version skew is not a protocol violation), and an
empty payload still means "the zero value". internal/daemon coverage
73.0% -> 78.4%.
