+++ internal/daemon/daemon.go
@@ -590,6 +590,12 @@ func (d *Daemon) serveControl(ctx context.Context, conn net.Conn) {
 		}
 		sendWorktrees(projectID, "list_worktrees_failed")
 	}
+	ops := controlOps{
+		writeJSON:      writeJSON,
+		sendError:      sendError,
+		sendWorktrees:  sendWorktrees,
+		finishMutation: finishMutation,
+	}
 	for {
 		ft, payload, err := wire.ReadFrame(conn)
 		if err != nil {
@@ -598,161 +604,184 @@ func (d *Daemon) serveControl(ctx context.Context, conn net.Conn) {
 			}
 			return
 		}
-		switch ft {
-		case wire.FrameShutdown:
-			// Return afterwards: the listener is about to close and
-			// this conn is going away with it. Nothing is written
-			// back — the client's proof of shutdown is the socket
-			// going quiet, not an ack it would have to trust.
-			d.Shutdown()
+		if d.handleControlFrame(ctx, ops, ft, payload) {
 			return
-		case wire.FrameListSessions:
-			_ = writeJSON(wire.FrameSessions, wire.SessionsResp{Sessions: d.reg.List()})
-		case wire.FrameCreateSession:
-			var spec wire.CreateSpec
-			if err := jsonUnmarshal(payload, &spec); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				// ErrNotFound here means the user killed the session
-				// while it was still being created — an ordinary
-				// outcome, not something to flash at them.
-				if _, err := d.reg.Create(ctx, spec); err != nil &&
-					!errors.Is(err, registry.ErrNotFound) {
-					sendError("create_failed", err.Error())
-				}
-			})
-		case wire.FrameKillSession:
-			var req wire.KillSessionReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				kill := d.reg.Kill
-				if req.RemoveWorktree {
-					kill = d.reg.KillAndRemoveWorktree
-				}
-				if err := kill(req.SessionID, req.Force); err != nil {
-					if errors.Is(err, registry.ErrWorktreeDirty) {
-						_ = writeJSON(wire.FrameError, wire.Error{
-							Code:      wire.ErrCodeWorktreeDirty,
-							Message:   "worktree has uncommitted changes",
-							SessionID: req.SessionID,
-						})
-					} else {
-						sendError("kill_failed", err.Error())
-					}
-				}
-			})
-		case wire.FrameRestartSession:
-			var req wire.RestartSessionReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				if err := d.reg.Restart(req.SessionID); err != nil {
-					sendError("restart_failed", err.Error())
-				}
-			})
-		case wire.FrameUpdateSession:
-			var req wire.UpdateSessionReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			if _, err := d.reg.Update(req); err != nil {
-				sendError("update_failed", err.Error())
-			}
-		case wire.FrameListProjects:
-			_ = writeJSON(wire.FrameProjects, wire.ProjectsResp{Projects: d.reg.ListProjects()})
-		case wire.FrameCreateProject:
-			var req wire.CreateProjectReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			if _, err := d.reg.CreateProject(req); err != nil {
-				sendError("create_project_failed", err.Error())
-			}
-		case wire.FrameKillProject:
-			var req wire.KillProjectReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				if err := d.reg.KillProject(req.ProjectID, req.KillSessions); err != nil {
-					sendError("kill_project_failed", err.Error())
-				}
-			})
-		case wire.FrameUpdateProject:
-			var req wire.UpdateProjectReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			if _, err := d.reg.UpdateProject(req); err != nil {
-				sendError("update_project_failed", err.Error())
-			}
-		case wire.FrameListWorktrees:
-			var req wire.ListWorktreesReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			// Off the read loop: the inventory shells out to git once
-			// per worktree, which on a big repo is slow enough to
-			// stall every other control request.
-			d.runOp(func() { sendWorktrees(req.ProjectID, "list_worktrees_failed") })
-		case wire.FrameRemoveWorktree:
-			var req wire.RemoveWorktreeReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				err := d.reg.RemoveWorktree(req.ProjectID, req.Path,
-					req.Force, req.DeleteBranch, req.DeleteRemote)
-				finishMutation(req.ProjectID, err, "remove_worktree_failed")
-			})
-		case wire.FrameCreateWorktree:
-			var req wire.CreateWorktreeReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				_, err := d.reg.CreateWorktreeForBranch(ctx, req.ProjectID, req.Branch)
-				finishMutation(req.ProjectID, err, "create_worktree_failed")
-			})
-		case wire.FrameDeleteBranch:
-			var req wire.DeleteBranchReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				err := d.reg.DeleteBranch(req.ProjectID, req.Branch,
-					req.Force, req.DeleteRemote)
-				finishMutation(req.ProjectID, err, "delete_branch_failed")
-			})
-		case wire.FrameRenameWorktree:
-			var req wire.RenameWorktreeReq
-			if err := jsonUnmarshal(payload, &req); err != nil {
-				sendError("bad_payload", err.Error())
-				continue
-			}
-			d.runOp(func() {
-				err := d.reg.RenameWorktree(req.ProjectID, req.Path, req.NewBranch)
-				finishMutation(req.ProjectID, err, "rename_worktree_failed")
-			})
-		default:
-			log.Printf("hived: unexpected control frame: %s", ft)
 		}
 	}
+}
+
+// controlOps bundles the per-connection reply helpers serveControl
+// builds over the socket's write mutex. Passing them as one value is
+// what lets the frame handlers live outside serveControl, which was
+// 274 lines of closure soup with a 14-case switch buried at the
+// bottom — untestable without a live socket, and the file's longest
+// function by a factor of two.
+type controlOps struct {
+	writeJSON      func(wire.FrameType, any) error
+	sendError      func(code, msg string)
+	sendWorktrees  func(projectID, failCode string)
+	finishMutation func(projectID string, err error, failCode string)
+}
+
+// decodeReq parses one request payload, answering `bad_payload` on the
+// connection and reporting false when it cannot. Twelve of the
+// fourteen control frames opened with this exact four-line prologue.
+func decodeReq[T any](payload []byte, sendError func(code, msg string)) (T, bool) {
+	var v T
+	if err := jsonUnmarshal(payload, &v); err != nil {
+		sendError("bad_payload", err.Error())
+		return v, false
+	}
+	return v, true
+}
+
+// handleControlFrame dispatches one control frame. It reports true
+// when the connection is finished (the client asked the daemon to shut
+// down), false to keep reading.
+func (d *Daemon) handleControlFrame(ctx context.Context, ops controlOps, ft wire.FrameType, payload []byte) bool {
+	switch ft {
+	case wire.FrameShutdown:
+		// Return afterwards: the listener is about to close and
+		// this conn is going away with it. Nothing is written
+		// back — the client's proof of shutdown is the socket
+		// going quiet, not an ack it would have to trust.
+		d.Shutdown()
+		return true
+	case wire.FrameListSessions:
+		_ = ops.writeJSON(wire.FrameSessions, wire.SessionsResp{Sessions: d.reg.List()})
+	case wire.FrameCreateSession:
+		spec, ok := decodeReq[wire.CreateSpec](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			// ErrNotFound here means the user killed the session
+			// while it was still being created — an ordinary
+			// outcome, not something to flash at them.
+			if _, err := d.reg.Create(ctx, spec); err != nil &&
+				!errors.Is(err, registry.ErrNotFound) {
+				ops.sendError("create_failed", err.Error())
+			}
+		})
+	case wire.FrameKillSession:
+		req, ok := decodeReq[wire.KillSessionReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			kill := d.reg.Kill
+			if req.RemoveWorktree {
+				kill = d.reg.KillAndRemoveWorktree
+			}
+			if err := kill(req.SessionID, req.Force); err != nil {
+				if errors.Is(err, registry.ErrWorktreeDirty) {
+					_ = ops.writeJSON(wire.FrameError, wire.Error{
+						Code:      wire.ErrCodeWorktreeDirty,
+						Message:   "worktree has uncommitted changes",
+						SessionID: req.SessionID,
+					})
+				} else {
+					ops.sendError("kill_failed", err.Error())
+				}
+			}
+		})
+	case wire.FrameRestartSession:
+		req, ok := decodeReq[wire.RestartSessionReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			if err := d.reg.Restart(req.SessionID); err != nil {
+				ops.sendError("restart_failed", err.Error())
+			}
+		})
+	case wire.FrameUpdateSession:
+		req, ok := decodeReq[wire.UpdateSessionReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		if _, err := d.reg.Update(req); err != nil {
+			ops.sendError("update_failed", err.Error())
+		}
+	case wire.FrameListProjects:
+		_ = ops.writeJSON(wire.FrameProjects, wire.ProjectsResp{Projects: d.reg.ListProjects()})
+	case wire.FrameCreateProject:
+		req, ok := decodeReq[wire.CreateProjectReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		if _, err := d.reg.CreateProject(req); err != nil {
+			ops.sendError("create_project_failed", err.Error())
+		}
+	case wire.FrameKillProject:
+		req, ok := decodeReq[wire.KillProjectReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			if err := d.reg.KillProject(req.ProjectID, req.KillSessions); err != nil {
+				ops.sendError("kill_project_failed", err.Error())
+			}
+		})
+	case wire.FrameUpdateProject:
+		req, ok := decodeReq[wire.UpdateProjectReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		if _, err := d.reg.UpdateProject(req); err != nil {
+			ops.sendError("update_project_failed", err.Error())
+		}
+	case wire.FrameListWorktrees:
+		req, ok := decodeReq[wire.ListWorktreesReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		// Off the read loop: the inventory shells out to git once
+		// per worktree, which on a big repo is slow enough to
+		// stall every other control request.
+		d.runOp(func() { ops.sendWorktrees(req.ProjectID, "list_worktrees_failed") })
+	case wire.FrameRemoveWorktree:
+		req, ok := decodeReq[wire.RemoveWorktreeReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			err := d.reg.RemoveWorktree(req.ProjectID, req.Path,
+				req.Force, req.DeleteBranch, req.DeleteRemote)
+			ops.finishMutation(req.ProjectID, err, "remove_worktree_failed")
+		})
+	case wire.FrameCreateWorktree:
+		req, ok := decodeReq[wire.CreateWorktreeReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			_, err := d.reg.CreateWorktreeForBranch(ctx, req.ProjectID, req.Branch)
+			ops.finishMutation(req.ProjectID, err, "create_worktree_failed")
+		})
+	case wire.FrameDeleteBranch:
+		req, ok := decodeReq[wire.DeleteBranchReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			err := d.reg.DeleteBranch(req.ProjectID, req.Branch,
+				req.Force, req.DeleteRemote)
+			ops.finishMutation(req.ProjectID, err, "delete_branch_failed")
+		})
+	case wire.FrameRenameWorktree:
+		req, ok := decodeReq[wire.RenameWorktreeReq](payload, ops.sendError)
+		if !ok {
+			return false
+		}
+		d.runOp(func() {
+			err := d.reg.RenameWorktree(req.ProjectID, req.Path, req.NewBranch)
+			ops.finishMutation(req.ProjectID, err, "rename_worktree_failed")
+		})
+	default:
+		log.Printf("hived: unexpected control frame: %s", ft)
+	}
+	return false
 }
 
 // serveAttach handles a session-attached connection.
